% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cosTransformPricer.R
\name{cosTransformPricer}
\alias{callCosCoeffs}
\alias{cosTransformPricer}
\alias{digitalCosCoeffs}
\title{Cosine Transform pricing}
\usage{
cosTransformPricer(strikeMat, mkt, N = 120, intLim,
  payCoeffFoo = callCosCoeffs, N.factors = 3, charFun = affineCF,
  preCalc = NULL, ...)

callCosCoeffs(intLim, Nterms = 120, strikeMat)

digitalCosCoeffs(intLim, Nterms, strikeMat)
}
\arguments{
\item{strikeMat}{array of size \code{TxKxS} of relative log-strikes}

\item{mkt}{data.frame with \code{T} rows and fields: r -- risk-free rate, q -- dividend yield, t -- option maturity.}

\item{N}{number of points of integration.}

\item{intLim}{\code{numeric(2)} integration limits in terms of log-strikes, see Fang and Oosterlee for some optimal choices.}

\item{payCoeffFoo}{function which accepts arguments \code{intLim, Nterms, strikeMat} and calculates the cosine expansion coefficents of a given derivative contract. This function determines what you are pricing. Use \code{callCosCoeffs} to price call (and put, via parity) options.}

\item{N.factors}{integer, number of stochastic volatility factors, argument for \code{charFun}. If your \code{charFun} doesn't accept such an argument, for example you're pricing in the Black-Scholes model, use \code{N.factors = 0}.}

\item{preCalc}{optional, array of size \code{NxTxKxS} with pre-calculated characteristic function values for integration. If you provide it, you can ommit \code{charFun} as it will not be called.}

\item{...}{arguments to charFun required for pricing (state variables, parameters, etc.)}
}
\value{
list of arrays of size \code{T x K x S} with relative prices of call options, put options, out of the money options.
}
\description{
Implementation of Fang and Oosterlee's (2008) Fourier-cosine expansion method of pricing European derivatives.
}
\details{
The integration method is described in detail in F. Fang and C.W. Oosterlee, ``A Novel Pricing Method for European Options Based on Fourier-Cosine Series Expansions'', accessible at http://ta.twi.tudelft.nl/mf/users/oosterle/oosterlee/COS.pdf.
}
\section{Functions}{
\itemize{
\item \code{callCosCoeffs}: 

\item \code{digitalCosCoeffs}: 
}}

