% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaussLaguerre.stateDeriv.R
\name{gaussLaguerre.stateDeriv}
\alias{gaussLaguerre.stateDeriv}
\title{Calculate derivatives of option prices with respect to state values, using a laguerre quadrature}
\usage{
gaussLaguerre.stateDeriv(params.Q, strikeMat, mkt, stateVec, rtol = 1e-13,
  alpha = 0, N = 64, preCalc = NULL, N.factors = 3)
}
\arguments{
\item{params.Q}{The Q-parameters of the stock-price process}

\item{strikeMat}{TxKxS array, containing the strikes at which the options are valued}

\item{mkt}{list describing the market structure (times to maturity, interest rate, dividend yield, current stock price)}

\item{stateVec}{SxN.factors matrix containing the states at which the pricing/derivatives are calculated}

\item{alpha}{parameter of the laguerre quadrature}

\item{N}{number of quadrature points to take}

\item{preCalc}{a list containing preCalculated values of the ODE solutions (useful if derivatives are reevaluated many times for different states, but the same parameter)}

\item{N.factors}{The number of stochastic factors we have}
}
\value{
A list containing elements call, put and otm, which in turn contain two lists of arrays of size TxKxS which show the sensitivity to state 1 or 2 at the given maturity,strike,state
}
\description{
Calculate derivatives of option prices with respect to state values, using a laguerre quadrature
}

