% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktUtilities.R
\name{makeFixedExpiryMktList}
\alias{makeFixedExpiryMktList}
\alias{makeFixedMaturityMktList}
\title{Utilities for creating mkt lists for simulating option prices}
\usage{
makeFixedExpiryMktList(num.days, t = 37/365, expiry.switch = 30,
  r = 0, p = 1, q = 0)

makeFixedMaturityMktList(num.days, r = 0, t = 30/365, q = 0, p = 1)
}
\arguments{
\item{num.days}{Length of list, how many days.}

\item{t}{vector of starting maturities}

\item{expiry.switch}{how often a new series of options is issued}

\item{r}{interest rate}

\item{p}{stock price}

\item{q}{dividend yield}

\item{num.days}{Length of list, how many days.}

\item{t}{vector of starting maturities}

\item{r}{interest rate}

\item{q}{dividend yield}

\item{p}{stock price}
}
\value{
A list mkt lists

A list containing lists svFast,svSlow and jmp
}
\description{
List with fixed expiry (changing maturity day by day)

List with fixed maturity each day
}
