% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_model_price_and_greeks.R
\name{option_price_with_greeks}
\alias{option_price_with_greeks}
\title{Option pricing and Greeks in affine models}
\usage{
option_price_with_greeks(strike_mat, mkt, type = c("call", "put", "otm"),
  N = 192L, intLim, payCoeffFoo = callCosCoeffs, params,
  N.factors = 1L, v.0, S.0, ode_sol = NULL, ...)
}
\arguments{
\item{strike_mat}{T x K x 1 array}

\item{mkt}{T x 1 data.frame}

\item{type}{call, put or otm options (character)}

\item{N}{integer num integration point see Fang and Oosterle 2008}

\item{intLim}{integration limits see Fang and Oosterle 2008}

\item{payCoeffFoo}{ibid}

\item{params}{Q parameter structure as in jumpDiffusionODEs in affineModelR}

\item{N.factors}{integer, number of factors in SV model}

\item{v.0}{matrix of states in pricing, 1 x N}

\item{S.0}{numeric, stock price at which options are evaluated}

\item{ode_sol}{array N x T x (1 + N.factors) of affine CF coefficients}

\item{...}{further arguments to affineCF}
}
\description{
Option pricing and Greeks in affine models
}
\details{
This works for a single time period!
}
