\name{transformOptionPricer-package}
\alias{transformOptionPricer-package}
\alias{transformOptionPricer}
\docType{package}
\title{
\packageTitle{transformOptionPricer}
}
\description{
\packageDescription{transformOptionPricer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{transformOptionPricer}
\packageIndices{transformOptionPricer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{transformOptionPricer}

Maintainer: \packageMaintainer{transformOptionPricer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
